chore(creatable-select-input): move to packages
