+++ packages/components/inputs/creatable-select-input/src/creatable-select-input.form.story.js
@@ -6,9 +6,9 @@ import { withKnobs, boolean } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
 import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
 import { ErrorMessage } from '@commercetools-uikit/messages';
-import Section from '../../../../.storybook/decorators/section';
-import FormikBox from '../../../../.storybook/decorators/formik-box';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import FormikBox from '../../../../../.storybook/decorators/formik-box';
+import Readme from '../README.md';
 import CreatableSelectInput from './creatable-select-input';
 
 const blueOption = { value: 'blue', label: 'Blue', color: '#0052CC' };
