+++ packages/components/inputs/creatable-select-input/src/creatable-select-input.js
@@ -15,8 +15,8 @@ import {
   TagRemove,
   DropdownIndicator,
 } from '@commercetools-uikit/select-utils';
-import createSelectStyles from '../../internals/create-select-styles';
-import messages from '../../internals/messages/select';
+import createSelectStyles from '../../../../../src/components/internals/create-select-styles';
+import messages from '../../../../../src/components/internals/messages/select';
 
 const useTheme = () => useContext(ThemeContext);
