+++ packages/components/inputs/creatable-select-input/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './creatable-select-input';
+
+export { default as version } from './version';
